REFACTOR: Put Graphviz support for Catlab `Graphs` in own submodule.
