+++ src/categorical_algebra/Graphs.jl
@@ -11,6 +11,7 @@ export AbstractGraph, Graph, AbstractSymmetricGraph, SymmetricGraph,
   gprops, vprops, eprops, get_gprop, get_vprop, get_eprop,
   set_gprop!, set_vprop!, set_eprop!, set_gprops!, set_vprops!, set_eprops!
 
+import Base: inv
 using Compat: isnothing
 import LightGraphs: SimpleGraph, SimpleDiGraph, nv, ne, src, dst,
   edges, vertices, has_edge, has_vertex, add_edge!, add_vertex!, add_vertices!,
